% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\name{applyCrudExpressionTree}
\alias{applyCrudExpressionTree}
\title{Apply access expression tree}
\usage{
applyCrudExpressionTree(inst, ...)
}
\arguments{
\item{inst}{Object for method dispatch.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See respective methods.
}
\description{
Applies a tree of access expressions.
See \code{\link[crudr]{applyCrudExpressionTree}}.
}
\details{
Known methods:

\itemize{
 \item{\code{\link[crudr]{applyCrudExpressionTree.default}}}
}
}
\examples{
inst = list(a = list(b = 1))

tree <- createCrudExpressionTree(inst, c("a", "b", "c"),
  value = 1, fail_safe = TRUE)

applyCrudExpressionTree(tree, "has")
applyCrudExpressionTree(tree, "read")
applyCrudExpressionTree(tree, "create")

inst
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{\link[crudr]{createCrudExpressionTree}}
}

