% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_adj_timezone.R
\name{cdr_adj_timezone}
\alias{cdr_adj_timezone}
\title{adjust for timezone treatment between databases}
\usage{
cdr_adj_timezone(conn_pool)
}
\arguments{
\item{conn_pool}{database connection of type pool}
}
\value{
'UTC' or empty string
}
\description{
adjust for timezone treatment between databases
}
\examples{
\dontrun{ cdr_adj_timezone(conn_pool) }

}
