% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_admin_key_ui.R
\name{cdr_admin_key_ui}
\alias{cdr_admin_key_ui}
\title{UI module interface button for row additions and/or row deletions in server module table}
\usage{
cdr_admin_key_ui(id)
}
\arguments{
\item{id}{your table name - namespace ID corresponding to the
'primary_tbl_name' in the database}
}
\value{
html (to be called in a shiny app only)
}
\description{
User interface portion of the server module Unique ID input textbox and button
button UI shows up for primary table if cell_edit_permission is set to T in server
}
\examples{
\dontrun{
ui <- shiny::fluidPage(crudr::cdr_admin_key_ui('primary_table_name'))
}
}
