% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_admin_tbl_ui.R
\name{cdr_admin_tbl_ui}
\alias{cdr_admin_tbl_ui}
\title{HTML to output a DT from the Server module}
\usage{
cdr_admin_tbl_ui(id, tbl = "db_tbl")
}
\arguments{
\item{id}{primary table name - namespace ID that matches the
table name of the primary table being administered in the database}

\item{tbl}{one of 'db_tbl' or 'chg_log_tbl' depending on which table
from server module you want to present}
}
\value{
a DTOutput table with loading spinner
}
\description{
A DT output for a namespace
}
\examples{
ui_html <- shiny::fluidPage(crudr:::cdr_admin_tbl_ui('iris', 'db_tbl'))

}
