% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_label_chg_log_activity.R
\name{cdr_attach_activity}
\alias{cdr_attach_activity}
\title{Determine the type of changes for the user change-log}
\usage{
cdr_attach_activity(chg_log, key_field = "UID")
}
\arguments{
\item{chg_log}{a change-log tibble pulled from the DB}

\item{key_field}{the field in the primary table that contains the unique-ID
for each observation}
}
\value{
the change-log tibble with activity type attached
}
\description{
Attach a column of data to the change-log describing the
type of change the user made. See: \code{\link[=cdr_label_chg_log_activity]{cdr_label_chg_log_activity()}}
}
\examples{
\dontrun{ cdr_attach_activity(chg_log, 'UID') }
}
