% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_chk_uniq_id.R
\name{cdr_chk_uniq_id}
\alias{cdr_chk_uniq_id}
\title{Checks availability of unique ID in DB}
\usage{
cdr_chk_uniq_id(db_tbl, input_uid, key_column, chk_for = c("create", "delete"))
}
\arguments{
\item{db_tbl}{tibble object: for primary table in server memory}

\item{input_uid}{char string: the unique ID you'd like to to add/remove for the
observation you're creating/deleting}

\item{key_column}{char string: that describes the column or field the UID is in}

\item{chk_for}{one of 'create' or 'delete': describing whether the user wants to
make a new UID or delete a row denoted by a UID. Create checking is slightly more
restrictive because it's capital insensitive.}
}
\value{
text to display in the UI
}
\description{
For a user to create or delete a new row observation, the ID for the unique ID column
must not already be used. Function checks user requested unique ID against
what's already in the DB.
}
