% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_chunk_tbl.R
\name{cdr_chunk_tbl}
\alias{cdr_chunk_tbl}
\title{break a dataframe into a list of dataframes}
\usage{
cdr_chunk_tbl(db_tbl, chunk_size = 10000)
}
\arguments{
\item{db_tbl}{a dataframe}

\item{chunk_size}{number of table elements to put into each chunk}
}
\value{
a list of dataframes
}
\description{
The number of dataframes in the list is
based on the number of elements specified for each chunk.
}
\examples{
chunked_df <- crudr:::cdr_chunk_tbl(iris, 20)
chunked_df <- crudr:::cdr_chunk_tbl(iris, 1)
chunked_df <- crudr:::cdr_chunk_tbl(iris, -1)
chunked_df <- crudr:::cdr_chunk_tbl(iris, 500)
chunked_df <- crudr:::cdr_chunk_tbl(iris, 1000)
chunked_df <- crudr:::cdr_chunk_tbl(as.matrix(iris), 0)
chunked_df <- crudr:::cdr_chunk_tbl(mtcars, 100)

}
