% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_coerce_value.R
\name{cdr_coerce_value}
\alias{cdr_coerce_value}
\title{change cell value using DT::coerceValue}
\usage{
cdr_coerce_value(input_val, old_mem_val)
}
\arguments{
\item{input_val}{cell_edit value}

\item{old_mem_val}{type from database table}
}
\value{
input_val value coerced to the correct df_tbl value
}
\description{
change cell value using DT::coerceValue
}
\examples{
\dontrun{
crudr::cdr_coerce_value(c('2023-03-17', '2023-03-17T14:28:24Z'), lubridate::now())
input_val <- c('true','t','truthy','tri',1,0,5,-5,'false','bla bla bla','a',' ','',
'na','none','no', 'yes', 'off', 'on', 'no', 'f','F','T', 'NO')
purrr::set_names(purrr::map(input_val, ~crudr::cdr_coerce_value(., TRUE)), input_val)
}
}
