% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_combine_prim_delt_tbls.R
\name{cdr_combine_prim_delt_tbls}
\alias{cdr_combine_prim_delt_tbls}
\title{Join primary and deltas tables}
\usage{
cdr_combine_prim_delt_tbls(db_tbl, db_tbl_deltas, key_col)
}
\arguments{
\item{db_tbl}{imported primary table}

\item{db_tbl_deltas}{imported deltas table}

\item{key_col}{the name of the unique ID column in the primary table}
}
\value{
the joined tables; all of db_tbl, and who-when information from the
deltas table (function also places the individual tables into output$ for
ui presentation)
}
\description{
join data from db_tbl and recent change values from db_tbl_deltas tables
(attaches most recent 'who' and 'when' form deltas table to primary table)
function used in 'mod_tbl_server' eventReactive
}
\examples{
\dontrun{
combine_prim_delt_tbls(db_tbl, db_tbl_deltas, key_col)
}
}
