% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_create_new_db_row.R
\name{cdr_create_new_db_row}
\alias{cdr_create_new_db_row}
\title{Appends new row to primary table}
\usage{
cdr_create_new_db_row(db_conn_pool, db_tbl_name, db_row)
}
\arguments{
\item{db_conn_pool}{pool object for the pool of connections to the specific db}

\item{db_tbl_name}{char string describing the specific table the new line is located in}

\item{db_row}{tibble of the row to append}
}
\value{
the uid created from the custom uid function
}
\description{
create new observation line in primary production database table and insert a unique
ID in the unique observation identifier line
Note: This function is only for the primary table. It is not used on the deltas table.
Note2: Function does not clean or control 'new_uid' inputs. Run your cleaning code before passing in 'new_uid'
}
