% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_create_new_db_tbls.R
\name{cdr_create_new_db_tbls}
\alias{cdr_create_new_db_tbls}
\title{Creates primary and deltas tables in a db}
\usage{
cdr_create_new_db_tbls(db_conn_pool, db_tbl)
}
\arguments{
\item{db_conn_pool}{a connection of class 'pool'}

\item{db_tbl}{a dataframe; the primary table to place into the database}
}
\value{
T or F; function cats statements to the console reporting success
or failure
}
\description{
instantiates an initial primary table into the database defined by
db_conn_pool; writes the initial primary table dataframe to the relational
database as well as instantiates the corresponding deltas metadata table for
change tracking Note: function should only be called once when setting up a
table to track within the db If you need to remove tables, these functions
can be helpful: pool::dbListTables(db_conn_pool)
pool::dbRemoveTable(db_conn_pool, 'some_table')
}
