% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_create_tbls_in_db.R
\name{cdr_create_tbls_in_db}
\alias{cdr_create_tbls_in_db}
\title{Creates primary and deltas tables in a db}
\usage{
cdr_create_tbls_in_db(db_conn_pool, db_tbl, key_field = NULL)
}
\arguments{
\item{db_conn_pool}{a database connection of class 'pool'}

\item{db_tbl}{a dataframe; the primary table to place into the database}

\item{key_field}{string: a unique ID column name in the data. If Null, creates a proxy}
}
\description{
creates an initial primary table into the database specified by
db_conn_pool; as well as the corresponding deltas metadata table for
the change tracking
}
\details{
Note: if you want to remove your table, these functions can be helpful:
pool::dbListTables(db_conn_pool)
pool::dbRemoveTable(db_conn_pool, 'some_table')
}
