% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_DB2RT_chg_log.R
\name{cdr_DB2RT_chg_log}
\alias{cdr_DB2RT_chg_log}
\title{Download database change log and send it to the module server}
\usage{
cdr_DB2RT_chg_log(conn_pool, chg_log_tbl_name)
}
\arguments{
\item{conn_pool}{pool connection object from package 'pool'}

\item{chg_log_tbl_name}{string: name of the database table you're managing with suffix "_DELTAS"}
}
\value{
a DT object to present the change log table
}
\description{
Download database change log and send it to the module server
}
\examples{
\dontrun{
con <- pool::dbPool(DBI::dbConnect(RSQLite::SQLite(), 'iris.db'))
cdr_make_db_tbls(con, iris)
cdr_DB2RT_primary(con, 'iris_DELTAS')
pool::dbRemoveTable(con,'iris')
pool::dbRemoveTable(con,'iris_DELTAS')
pool::poolClose(con)
}

}
