% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_DB2RT_primary.R
\name{cdr_DB2RT_primary}
\alias{cdr_DB2RT_primary}
\title{Download database table and sync it to user interface}
\usage{
cdr_DB2RT_primary(conn_pool, db_tbl_name, key_col)
}
\arguments{
\item{conn_pool}{pool connection object from package 'pool'}

\item{db_tbl_name}{string: name of the specific table to update or \code{\link[=cdr_id]{cdr_id()}} object}

\item{key_col}{string: the field that contains your unique ID key for each row}
}
\value{
a DT object to present the primary table
}
\description{
Download database table and sync it to user interface
}
\examples{
\dontrun{
con <- pool::dbPool(DBI::dbConnect(RSQLite::SQLite(), 'iris.db'))
cdr_make_db_tbls(con, iris)
cdr_DB2RT_primary(con, 'iris')
pool::dbRemoveTable(con,'iris')
pool::dbRemoveTable(con,'iris_DELTAS')
pool::poolClose(con)
}

}
