% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_deliver_admin_tbl.R
\name{cdr_deliver_admin_tbl}
\alias{cdr_deliver_admin_tbl}
\title{Shiny UI HTML for the administrator table and}
\usage{
cdr_deliver_admin_tbl(db_tbl_name = "")
}
\arguments{
\item{db_tbl_name}{string: name of the primary database table being managed (for namespace id)}
}
\value{
the rendered Datatable interface
}
\description{
Shiny UI HTML for the administrator table and
}
\examples{
ui_html <- cdr_deliver_admin_tbl('iris')

}
