% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_deliver_chg_log_tbl.R
\name{cdr_deliver_chg_log_tbl}
\alias{cdr_deliver_chg_log_tbl}
\title{gets the change history table for the Shiny UI}
\usage{
cdr_deliver_chg_log_tbl(db_tbl_name = "")
}
\arguments{
\item{db_tbl_name}{string: name of the primary database table being managed}
}
\value{
the rendered Datatable interface
}
\description{
gets the change history table for the Shiny UI
}
\examples{
\dontrun{ crudr::cdr_deliver_chg_log_tbl('iris') }
}
