% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_id.R
\name{cdr_id}
\alias{cdr_id}
\title{A drop-in replacement for \code{\link[DBI:Id]{DBI::Id()}}, but with ordered output}
\usage{
cdr_id(...)
}
\arguments{
\item{...}{DB table index parameters, i.e. strings with with labels
table, schema, catalog, cluster, database}
}
\value{
equivalent \code{\link[DBI:Id]{DBI::Id()}} object
}
\description{
A drop-in replacement for \code{\link[DBI:Id]{DBI::Id()}}, but with ordered output
}
\examples{
cdr_id(table = 'my_table', schema = 'a_schema', some_new_spec = 'in_case_you_need_it')
cdr_id(table = 'table_last', schema = 'schema_3rd', cluster = 'clus_1st', catalog = 'cat_2nd')

}
