% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_id2sql.R
\name{cdr_id2sql}
\alias{cdr_id2sql}
\title{Convert a name or DBI::Id to SQL}
\usage{
cdr_id2sql(db_tbl_name)
}
\arguments{
\item{db_tbl_name}{db name as text or \code{\link[=cdr_id]{cdr_id()}} object}
}
\value{
SQL table object
}
\description{
Convert a name or DBI::Id to SQL
}
\examples{
crudr:::cdr_id2sql('hello')
crudr:::cdr_id2sql(cdr_id(table = 'hello', schema = 'blue'))
crudr:::cdr_id(table = 'hello', schema = 'blue')
crudr:::cdr_id2sql(cdr_id(schema = 'blue', table = 'hello'))
crudr:::cdr_id(schema = 'blue', table = 'hello')

}
