% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_impart_chg_log_tbl.R
\name{cdr_impart_chg_log_tbl}
\alias{cdr_impart_chg_log_tbl}
\title{Download database change log and sync it to user interface}
\usage{
cdr_impart_chg_log_tbl(
  db_conn_pool,
  chg_log_tbl_name = crudr::cdr_name_delta_tbl(db_tbl_name)
)
}
\arguments{
\item{db_conn_pool}{pool connection object from package 'pool'}

\item{chg_log_tbl_name}{string: name of the database table you're managing with suffix "_DELTAS"}
}
\value{
a DT object to present the change log table
}
\description{
Download database change log and sync it to user interface
}
\examples{
\dontrun{cdr_impart_primary_tbl(con, 'IRIS_DELTAS')}
}
