% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_impart_primary_tbl.R
\name{cdr_impart_primary_tbl}
\alias{cdr_impart_primary_tbl}
\title{Download database table and sync it to user interface}
\usage{
cdr_impart_primary_tbl(
  db_conn_pool,
  db_tbl_name,
  key_col,
  cell_edit_permission,
  lock_fields = c()
)
}
\arguments{
\item{db_conn_pool}{pool connection object from package 'pool'}

\item{db_tbl_name}{string: name of the database table you're managing}

\item{key_col}{string: the field that contains your unique ID per row}

\item{cell_edit_permission}{bool: do you want to allow the admin table cells to be edited}

\item{lock_fields}{strings: column names you want to lock up e.g. c('Species','Petal.Width')}
}
\value{
a DT object to present the primary table
}
\description{
Download database table and sync it to user interface
}
\examples{
\dontrun{cdr_impart_primary_tbl(con, 'IRIS', 'UID', TRUE, 'Species')}
}
