% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_join_tbls.R
\name{cdr_join_tbls}
\alias{cdr_join_tbls}
\title{Join primary and deltas tables}
\usage{
cdr_join_tbls(db_tbl, chg_log_tbl, key_col)
}
\arguments{
\item{db_tbl}{imported primary table}

\item{chg_log_tbl}{imported deltas table}

\item{key_col}{the name of the unique ID column in the primary table}
}
\value{
the joined tables; all of db_tbl, and who-when information from the
deltas table (function also places the individual tables into output$ for
ui presentation)
}
\description{
join data from db_tbl and recent change values from chg_log_tbl tables
(attaches most recent 'who' and 'when' form deltas table to primary table)
function used in 'mod_tbl_server' eventReactive
}
\examples{
\dontrun{
combine_prim_delt_tbls(db_tbl, chg_log_tbl, key_col)
}
}
