% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_js_edit_ctrl.R
\name{cdr_js_edit_ctrl}
\alias{cdr_js_edit_ctrl}
\title{Allows key use to navigate table and edit cells}
\usage{
cdr_js_edit_ctrl()
}
\value{
javascript callback for DT::datatable()
}
\description{
Allows key use to navigate table and edit cells
}
\examples{
crudr::cdr_js_edit_ctrl()
}
