% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_label_chg_log_activity.R
\name{cdr_label_chg_log_activity}
\alias{cdr_label_chg_log_activity}
\title{Create change types for the change-log table}
\usage{
cdr_label_chg_log_activity(chg_log, key_field = "UID")
}
\arguments{
\item{chg_log}{a change-log tibble pulled from the DB}

\item{key_field}{the field in the primary table that contains the unique-ID
for each observation}
}
\value{
the change-log tibble with activity type attached
}
\description{
Attach a column of data to the change-log describing the
type of user change, i.e.
\itemize{
\item 'created' - user created a new ID
\item 'deleted' user deleted an ID and all associated row elements
\item 'updated' user made a change to a field of an existing ID
}
}
\examples{
\dontrun{
  cdr_label_chg_log_activity(chg_log, 'UID')
}

}
