% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_mod_btn_ui.r
\name{cdr_mod_btn_ui}
\alias{cdr_mod_btn_ui}
\title{UI module interface for row additions in server module table}
\usage{
cdr_mod_btn_ui(id)
}
\arguments{
\item{id}{primary table name - namespace ID corresponding to the
'primary_tbl_name' in the database}
}
\value{
html (to be called in a shiny app only)
}
\description{
User interface portion of the server module UID input textbox and submit
button UI shows up for primary table if open_sesame is set to T in server
}
\examples{
\dontrun{
ui <- fluidPage(mod_tbl_ui('primary_table_name'))
}
}
