% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_mod_tbl_ui.r
\name{cdr_mod_tbl_ui}
\alias{cdr_mod_tbl_ui}
\title{UI module for a DT from the Server module}
\usage{
cdr_mod_tbl_ui(id, tbl_type = "db_tbl")
}
\arguments{
\item{id}{primary table name - namespace ID corresponding to the
'primary_tbl_name' in the database}

\item{tbl_type}{one of 'db_tbl' or 'db_tbl_deltas' depending on which table
from mod_tbl_server you want to present}
}
\value{
html (to be called in a shiny app only)
}
\description{
User interface portion of the crudr module. Presents an ajax datatable in a
shiny UI. Just a DT output wrapped in a namespace
}
\examples{
\dontrun{
ui <- fluidPage(mod_tbl_ui('primary_table_name'))
}
}
