% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_reconstruct_past_tbl.R
\name{cdr_reconstruct_past_tbl}
\alias{cdr_reconstruct_past_tbl}
\title{Build the admin table as it was at a point in the past}
\usage{
cdr_reconstruct_past_tbl(
  as_of = lubridate::now() - lubridate::weeks(x = 4),
  conn_pool,
  db_tbl_name,
  chg_log_suffix = "_DELTAS",
  key_field = "UID",
  ...
)
}
\arguments{
\item{as_of}{datetime object in the past, e.g. lubridate::now()-lubridate::weeks(4)}

\item{conn_pool}{pool or DBI connection}

\item{db_tbl_name}{primary table name}

\item{chg_log_suffix}{your table name extension specifying the corresponding
change-log table, e.g. '_deltas'}

\item{key_field}{column in the primary table that acts as the unique row identifier}

\item{...}{other connection parameters like \code{schema = public}}
}
\value{
list with primary and change-log tables reconstructed as they were at
a time in the past
}
\description{
Build the admin table as it was at a point in the past
}
\examples{
\dontrun{
conn_pool <- pool::dbPool(
  DBI::dbConnect( drv = RPostgreSQL::PostgreSQL(),
                  dbname = "test",
                  host = "localhost",
                  port = "5432",
                  user = Sys.getenv('postgres_username'),
                  password = Sys.getenv('postgres_password'))
)
cdr_reconstruct_past_tbl(conn_pool = conn_pool, db_tbl_name = 'mtcars', schema = "test")
cdr_reconstruct_past_tbl(conn_pool = conn_pool, db_tbl_name = 'mtcars', schema = "test")
pool::poolClose(conn_pool)
}

}
