% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_remove_tbl.R
\name{cdr_remove_tbl}
\alias{cdr_remove_tbl}
\title{removes a specified table by deleting the table or by removing just the table data (truncating)}
\usage{
cdr_remove_tbl(
  db_conn_pool,
  db_tbl_name = NULL,
  removal = c("delete", "truncate")
)
}
\arguments{
\item{db_conn_pool}{pool connection object: the pool of connections established by the session}

\item{db_tbl_name}{string: name of the specific table the value to update is located in}

\item{removal}{string: one of 'delete' or 'truncate'}
}
\value{
Boolean describing success
}
\description{
removes a specified table by deleting the table or by removing just the table data (truncating)
}
\examples{
\dontrun{cdr_remove_tbl(db_conn_pool = pool_connection_object, db_tbl_name = iris, removal = 'truncate')}

}
