% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_render_new_row_ui.r
\name{cdr_render_new_row_ui}
\alias{cdr_render_new_row_ui}
\title{new observation row addition UI}
\usage{
cdr_render_new_row_ui(notes_txt = "", id, open_sesame)
}
\arguments{
\item{notes_txt}{message to send to the UI edit page}

\item{id}{namespace id (corresponds to DB name)}

\item{open_sesame}{whether user has permission to edit page}
}
\value{
ui output component
}
\description{
new observation row addition UI
}
\examples{
cdr_render_new_row_ui('hello', 'iris', TRUE)
}
