% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_reset_db.R
\name{cdr_reset_db}
\alias{cdr_reset_db}
\title{removes the tables written to a db and closes the connection}
\usage{
cdr_reset_db(conn_pool, db_tbl_name = NULL)
}
\arguments{
\item{conn_pool}{pool connection object: the pool of connections established by the session}

\item{db_tbl_name}{string: name of the specific table the value to update is located in}
}
\description{
generally for testing
}
\examples{
\dontrun{ reset_db(con, iris) }
}
