% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_row_editor_html.R
\name{cdr_row_editor_html}
\alias{cdr_row_editor_html}
\title{The html for the UI to add and delete new rows}
\usage{
cdr_row_editor_html(
  notes_txt = "",
  id,
  add_row_permission = F,
  del_row_permission = F
)
}
\arguments{
\item{notes_txt}{message to send to the UI edit page}

\item{id}{namespace id - corresponds to table name}

\item{add_row_permission}{T or F: whether user has permission to add a new observation, i.e. the Unique ID for a row}

\item{del_row_permission}{T or F: whether user has permission to delete a new observation, i.e. the Unique ID for a row}
}
\value{
ui output component
}
\description{
Responsive to given permissions
}
\examples{
crudr:::cdr_row_editor_html('hello', 'iris', TRUE, TRUE)
}
