% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_RT2DT_chg_log.R
\name{cdr_RT2DT_chg_log}
\alias{cdr_RT2DT_chg_log}
\title{Format the change log R table into a DT}
\usage{
cdr_RT2DT_chg_log(chg_log_tbl)
}
\arguments{
\item{chg_log_tbl}{change log table imparted from the DB}
}
\value{
a DT object
}
\description{
Format the change log R table into a DT
}
\examples{
\dontrun{
crudr:::cdr_RT2DT_chg_log(iris)
}
}
