% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_RT2DT_primary.R
\name{cdr_RT2DT_primary}
\alias{cdr_RT2DT_primary}
\title{Format the primary R table into a DT}
\usage{
cdr_RT2DT_primary(db_tbl, cell_edit_permission, lock_fields = c())
}
\arguments{
\item{db_tbl}{change log table imparted from the DB}

\item{cell_edit_permission}{bool: do you want to allow the admin table cells to be edited}

\item{lock_fields}{strings: column names you want to lock up e.g. c('Species','Petal.Width')}
}
\value{
a DT object to present the primary table
}
\description{
Format the primary R table into a DT
}
\examples{
\dontrun{
crudr:::cdr_RT2DT_primary(iris, cell_edit_permission = T)
}

}
