% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_std_pad.R
\name{cdr_std_pad}
\alias{cdr_std_pad}
\title{Integer sequence with preceding zeroes}
\usage{
cdr_std_pad(n)
}
\arguments{
\item{n}{length of the sequence}
}
\value{
a character vector
}
\description{
Integer sequence with preceding zeroes
}
\examples{
crudr:::cdr_std_pad(3)
crudr:::cdr_std_pad(10)

}
