% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_tbl2ansi_format.R
\name{cdr_tbl2ansi_format}
\alias{cdr_tbl2ansi_format}
\title{Prep dataframe columns to become SQL insert stmts}
\usage{
cdr_tbl2ansi_format(db_tbl, tzone = "UTC")
}
\arguments{
\item{db_tbl}{a dataframe}

\item{tzone}{the database default timezone}
}
\value{
a tibble
}
\description{
Prep dataframe columns to become SQL insert stmts
}
\examples{
example_df <- dplyr::mutate(iris, bool = Species == 'setosa', day = Sys.Date(), test = Sys.time())
adj_df <- crudr:::cdr_tbl2ansi_format(example_df)
adj_df <- crudr:::cdr_tbl2ansi_format(example_df, Sys.timezone())

}
