% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_update_deltas_tbl.R
\name{cdr_update_deltas_tbl}
\alias{cdr_update_deltas_tbl}
\title{Populates deltas table when user updates primary table}
\usage{
cdr_update_deltas_tbl(
  db_conn_pool,
  db_tbl_name,
  old_value,
  update_value,
  value_colname,
  value_rowuid,
  who = Sys.info()[["user"]]
)
}
\arguments{
\item{db_conn_pool}{see update_primary_tbl() definitions}

\item{db_tbl_name}{see update_primary_tbl() definitions}

\item{old_value}{the value that was overwritten in the primary table}

\item{update_value}{the value that overwrote 'old_value' in the primary table}

\item{value_colname}{see update_primary_tbl() definitions}

\item{value_rowuid}{see update_primary_tbl() definitions}

\item{who}{string: the username responsible for the change}
}
\value{
an in-memory copy of the data just appended to the tracking table
}
\description{
Connects to and posts who-what-when data to a change tracking table
Function is called when a user updates the data in a primary table
}
