% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_update_primary_tbl.R
\name{cdr_update_primary_tbl}
\alias{cdr_update_primary_tbl}
\title{updates the primary db table by specifying value to change,
corresponding db table, unique row ID, and column}
\usage{
cdr_update_primary_tbl(
  db_conn_pool = db_conn_pool,
  db_tbl_name = NULL,
  update_value = NULL,
  value_colname = NULL,
  value_rowuid = NULL,
  value_rowuid_colname = NULL
)
}
\arguments{
\item{db_conn_pool}{pool connection object: the pool of connections established by the session}

\item{db_tbl_name}{string: name of the specific table the value to update is located in}

\item{update_value}{number or string: the value to update already in the correct data type}

\item{value_colname}{string: the specific column name where the value to update is located}

\item{value_rowuid}{number or string: the specific row unique ID that corresponds to the row
where the value to update is located (key)}

\item{value_rowuid_colname}{string: the name of the column with the unique ID (key column)}
}
\value{
TRUE
}
\description{
updates the primary db table by specifying value to change,
corresponding db table, unique row ID, and column
}
