% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique_id.r
\name{check_unique_id}
\alias{check_unique_id}
\title{Checks availability of unique ID in DB}
\usage{
check_unique_id(db_conn_pool, db_tbl_name, new_uid, uid_column_name)
}
\arguments{
\item{db_conn_pool}{pool object for the pool of connections to the specific db}

\item{db_tbl_name}{char string describing the specific table the new line is located in}

\item{new_uid}{numeric or string: the new unique ID you'd like to use for the new observation you're creating; input type must match the db type}

\item{uid_column_name}{character string describing the unique ID column}
}
\value{
the uid created from the custom uid function
}
\description{
For a user to create a new row observation, the ID for the unique ID column must not already be used.
Function checks user requested unique ID against what's already in the DB
}
