% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{createCrudExpressionPlain.default}
\alias{createCrudExpressionPlain.default}
\title{Create access expression}
\usage{
\method{createCrudExpressionPlain}{default}(inst, id = character(),
  value = NULL, name = "inst", ...)
}
\arguments{
\item{inst}{\code{\link[base]{list}}.
TODO}

\item{id}{\code{\link[base]{character}}.
TODO}

\item{value}{\code{ANY} (\code{\link[methods]{BasicClasses}}).
Optional assignment value.}

\item{name}{\code{\link[base]{character}}.
Name of the actual object in the expression that values are extracted from
or assigned to.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link[base]{expression}}.
}
\description{
Default method of \code{\link[crudr]{createCrudExpressionPlain}}.
}
\details{
TODO
}
\examples{
inst <- list(a = list(b = list(c = 10)))

(expr <- createCrudExpressionPlain(inst, c("a", "b")))
eval(expr)

(expr <- createCrudExpressionPlain(inst, c("a", "b"), 1))
eval(expr)
inst
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{
 \link[crudr]{createCrudExpressionPlain},
 \link[crudr]{createCrudExpression}
}
}

