% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\name{createCrudExpressionPlain}
\alias{createCrudExpressionPlain}
\title{Create access expression}
\usage{
createCrudExpressionPlain(inst, ...)
}
\arguments{
\item{inst}{Object for method dispatch.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See respective methods.
}
\description{
Plain and thus faster version of
\code{\link[crudr]{createCrudExpression}} for
scenarios where computational costs really matter.
}
\details{
Known methods:

\itemize{
 \item{\code{\link[crudr]{createCrudExpressionPlain.default}}}
}

For a more comprehensive but slower version see:
\code{\link[crudr]{createCrudExpression}}
}
\examples{
inst <- list(a = list(b = list(c = 10)))

(expr <- createCrudExpressionPlain(inst, c("a", "b")))
eval(expr)

(expr <- createCrudExpressionPlain(inst, c("a", "b"), 1))
eval(expr)
inst
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{\link[base]{Extract}, \link[base]{assignOps}}
}

