% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{createCrudExpressionTree.default}
\alias{createCrudExpressionTree.default}
\title{Create access expression tree}
\usage{
\method{createCrudExpressionTree}{default}(inst, id = character(),
  value = NULL, name_obj = "inst", name_value = "value",
  as_name_obj = TRUE, as_name_value = FALSE, fail_safe = FALSE, ...)
}
\arguments{
\item{inst}{\code{\link[base]{list}}.
TODO}

\item{id}{\code{\link[base]{character}}.
TODO}

\item{value}{\code{ANY} (\code{\link[methods]{BasicClasses}}).
Optional assignment value.}

\item{name_obj}{\code{\link[base]{character}}.
Name of the actual object in the expression that values are extracted from
or assigned to.}

\item{name_value}{\code{\link[base]{character}}.
TODO}

\item{as_name_obj}{\code{\link[base]{logical}}.
TODO}

\item{as_name_value}{\code{\link[base]{logical}}.
TODO}

\item{fail_safe}{\code{\link[base]{logical}}.
Wrap with \code{\link[base]{try}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link[base]{expression}}.
}
\description{
Default method of \code{\link[crudr]{createCrudExpressionTree}}.
}
\details{
TODO
}
\examples{

# Example object ----------------------------------------------------------

inst = list()

# No assignment value -----------------------------------------------------

createCrudExpressionTree(inst, c("a", "b", "c"))
createCrudExpressionTree(inst, c("a", "b", "c"), name_obj = "x")
createCrudExpressionTree(inst, c("a", "b", "c"), as_name_obj = FALSE)

createCrudExpressionTree(inst, c("a", "b", "c"), fail_safe = TRUE)

# Assignment value --------------------------------------------------------

createCrudExpressionTree(inst, c("a", "b", "c"), value = 1)
createCrudExpressionTree(inst, c("a", "b", "c"), value = 1,
  name_value = "y", as_name_value = TRUE)

createCrudExpressionTree(inst, c("a", "b", "c"), value = 1, fail_safe = TRUE)

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{
 \link[crudr]{createCrudExpressionTree}
}
}

