% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crud.R
\name{createCrudExpressionTree}
\alias{createCrudExpressionTree}
\title{Create access expression tree}
\usage{
createCrudExpressionTree(inst, ...)
}
\arguments{
\item{inst}{Object for method dispatch.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See respective methods.
}
\description{
Creates a tree of access expression.
See \code{\link[crudr]{createCrudExpression}}.
}
\details{
Known methods:

\itemize{
 \item{\code{\link[crudr]{createCrudExpressionTree.default}}}
}
}
\examples{

# Example object ----------------------------------------------------------

inst = list()

# No assignment value -----------------------------------------------------

tree <- createCrudExpressionTree(inst, c("a", "b", "c"))
sapply(tree, names)

createCrudExpressionTree(inst, c("a", "b", "c"), name_obj = "x")
createCrudExpressionTree(inst, c("a", "b", "c"), as_name_obj = FALSE)

createCrudExpressionTree(inst, c("a", "b", "c"), fail_safe = TRUE)

# Assignment value --------------------------------------------------------

createCrudExpressionTree(inst, c("a", "b", "c"), value = 1)
createCrudExpressionTree(inst, c("a", "b", "c"), value = 1,
  name_value = "y", as_name_value = TRUE)

createCrudExpressionTree(inst, c("a", "b", "c"), value = 1, fail_safe = TRUE)

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{\link[base]{Extract}, \link[base]{assignOps}}
}

