% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation_classes.R
\docType{data}
\name{Crud.Settings}
\alias{Crud.Settings}
\title{Base class for CRUD Operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Crud.Settings
}
\description{
TODO
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{main}}{\code{\link{list}}
Main information}
}}
\section{Getters/setters}{


\itemize{
 \item{See interface} {
   \code{\link[crudr]{ICrud}}
 }
}
}

\section{Public methods}{


\itemize{
 \item{See interface} {
   \code{\link[crudr]{ICrud}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\keyword{datasets}

