% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crudr-package.r
\docType{package}
\name{crudr-package}
\alias{crudr}
\alias{crudr-package}
\title{Interfaces for CRUD operations}
\description{
Provides generic interfaces that can be systematically
implemented for all contexts that match a CRUD paradigm (create, read,
update, delete).
}
\details{
Details
}
\section{Classes and constructors}{


\itemize{
   \item{\code{\link[crudr]{crudr}}: }{

     TODO
   }
}
}

\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[crudr]{crudr}}: }{

     TODO
   }
}
}

\section{Disclaimer}{

\strong{This package is work-in-progress and is thus likely to change quite
substantially until it reaches a more mature state.}
}

\section{Aknowledgements}{

This package was inspired by the following resources:

\itemize{
   \item{Package \code{settings}: } {

     \url{https://github.com/markvanderloo/settings}
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
See \code{vignette("introduction", package = "crudr")}.
}

