% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{handleThreedots}
\alias{handleThreedots}
\title{Handle threedots input}
\usage{
handleThreedots(...)
}
\arguments{
\item{...}{Flexible input that should be passed along to downstream
functions.}
}
\value{
\code{\link[base]{list}}
}
\description{
Handles threedots input for downstream functions whose main input is
provided via \code{...}.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{\link[base]{Reserved}, \link[methods]{dotsMethods}}
}

