% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_classes.R
\docType{data}
\name{ICrud}
\alias{ICrud}
\title{Interface Class for CRUD Operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ICrud
}
\description{
Defines the interface for CRUD operations.
}
\section{Details on interface term}{

The term \emph{interface} is used in a looser context than in more
rigid OOP contexts such as \emph{C#} or the like
}

\section{Getters/setters}{


\itemize{
 \item{\code{getMain}} {

   Get main field value
 }
 \item{\code{setMain}} {

   Set main field value
 }
}
}

\section{Public methods}{


\itemize{
 \item{init} {

   Has
 }
 \item{has} {

   Has
 }
 \item{create} {

   Create
 }
 \item{read} {

   Read
 }
 \item{update} {

   Update
 }
 \item{delete} {

   Delete
 }
 \item{reset} {

   Reset
 }
}
}

\section{Known interface implementations}{


\itemize{
 \item{\code{\link[crudr]{Crud}}}
 \item{\code{\link[crudr]{Crud.Settings}}}
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\keyword{datasets}

