% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_tbl_server.r
\name{mod_tbl_server}
\alias{mod_tbl_server}
\title{shiny server module gets database data from db specified by db_conn_pool connection and the db table name specified in 'id',
server module manages and syncs changes between the UI DT, in-server-memory tbl, and backend db}
\usage{
mod_tbl_server(id, key_col, db_conn_pool, who)
}
\arguments{
\item{id}{primary table name - namespace ID corresponding to the 'primary_tbl_name' in the database}

\item{key_col}{name of the unique ID column in the db table (table must have a unique ID column with unique IDs)}

\item{db_conn_pool}{db connection from package 'pool'}
}
\value{
returns DT reactive tables to the shiny ui environment
}
\description{
shiny server module gets database data from db specified by db_conn_pool connection and the db table name specified in 'id',
server module manages and syncs changes between the UI DT, in-server-memory tbl, and backend db
}
\examples{
\dontrun{
db_conn_pool <- easydbconn::get_db_con(friendly_db_name) # ensure db is already setup with crudr::create_new_db_tbls()
server <- function(input, output, session){mod_tbl_server('primary_tbl_name', 'key_col_name', db_conn_pool)}
shinyApp(ui, server)
}
}
