% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_mod_tbl_ui.r
\name{mod_tbl_ui}
\alias{mod_tbl_ui}
\title{User interface portion of the crudr module
Presents an ajax datatable in a shiny UI
Used exclusively on 'mod_tbl_server' object}
\usage{
mod_tbl_ui(id, tbl_type = "db_tbl")
}
\arguments{
\item{id}{primary table name - namespace ID corresponding to the 'primary_tbl_name' in the database}

\item{tbl_type}{one of 'db_tbl' or 'db_tbl_deltas' depending on which table from mod_tbl_server you want to present}
}
\value{
html (to be called in a shiny app only)
}
\description{
User interface portion of the crudr module
Presents an ajax datatable in a shiny UI
Used exclusively on 'mod_tbl_server' object
}
\examples{
\dontrun{
ui <- fluidPage(mod_tbl_ui('primary_table_name'))
}
}
