% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threedots.R
\name{withHandledThreedots}
\alias{withHandledThreedots}
\title{Call function with handle threedots input}
\usage{
withHandledThreedots(..., fun)
}
\arguments{
\item{...}{Flexible input that should be passed along to downstream
function \code{fun}.}

\item{fun}{\code{\link[base]{function}}.
Function to be called with handled threedots input.}
}
\value{
Whatever the function \code{fun} returns.
}
\description{
Calls a function whose main input is provided via \code{...} and passes
a handled version of the original \code{...} as returend by
\code{\link[crudr]{handleThreedots}}.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/crudr}
}
\seealso{
\code{\link[crudr]{handleThreedots}}
}

