% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_admin_tbl_ui.R
\name{cdr_admin_tbl_ui}
\alias{cdr_admin_tbl_ui}
\title{UI module to output a DT from the Server module}
\usage{
cdr_admin_tbl_ui(id, tbl = "db_tbl")
}
\arguments{
\item{id}{primary table name - namespace ID that matches the
table name of the primary table being administered in the database}

\item{tbl}{one of 'db_tbl' or 'chg_log_tbl' depending on which table
from server module you want to present}
}
\value{
a DTOutput table with loading spinner
}
\description{
User interface portion of the crudr module. Presents an ajax datatable in a
shiny UI. Just a DT output wrapped in a namespace
}
\examples{
\dontrun{
ui <- fluidPage(mod_tbl_ui('primary_table_name'))
}
}
