% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_append_tbl.R
\name{cdr_append_tbl}
\alias{cdr_append_tbl}
\title{Appends data from a dataframe into an existing database table with the same structure}
\usage{
cdr_append_tbl(conn_pool, db_tbl, db_tbl_name = NULL, chunk_size = 10000)
}
\arguments{
\item{conn_pool}{a database connection of class 'pool'}

\item{db_tbl}{the dataframe to append to the database table}

\item{db_tbl_name}{the name of the database table if different from the name of the dataframe passed to 'db_tbl'}

\item{chunk_size}{the maximum number of elements you want to pass to the db in one go
(i.e. if you have a huge dataframe with 100 columns and 100,000 rows, a
chunk_size of 10,000 elements would split the dataframe into 1000 groups and
appends each successively.)}
}
\description{
Appends data from a dataframe into an existing database table with the same structure
}
\examples{
\dontrun{
con <- pool::dbPool(DBI::dbConnect(RSQLite::SQLite(), 'test.db'))
iris_tbl <- dplyr::mutate(iris, bool = Species == 'setosa', day = Sys.Date(), test = Sys.time())
pool::dbCreateTable(con,  'iris_tbl', iris_tbl)
crudr::cdr_append_tbl(con, iris_tbl)
dplyr::tbl(con, 'iris_tbl')
pool::dbRemoveTable(con,'iris_tbl')
pool::poolClose(con)
}

}
