% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_coerce_value.R
\name{cdr_coerce_value}
\alias{cdr_coerce_value}
\title{change cell_edit value to match database value}
\usage{
cdr_coerce_value(input, old_mem_val)
}
\arguments{
\item{input}{cell_edit value}

\item{old_mem_val}{type from database table}
}
\value{
input value coerced to the correct df_tbl value
}
\description{
change cell_edit value to match database value
}
\examples{
{
input <- c('true','t','truthy','tri',1,0,'false','bla bla bla','a',' ','', 'na','none','no')
purrr::set_names(purrr::map(input, ~cdr_coerce_value(., TRUE)), input)
}
}
