% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_delete_row_in_db.R
\name{cdr_delete_row_in_db}
\alias{cdr_delete_row_in_db}
\title{Removes specified row from db table}
\usage{
cdr_delete_row_in_db(
  db_conn_pool = db_conn_pool,
  db_tbl_name = NULL,
  value_rowuid = NULL,
  key_column = NULL
)
}
\arguments{
\item{db_conn_pool}{pool connection object: the pool of connections established by the session}

\item{db_tbl_name}{string: name of the specific table the value to update is located in}

\item{value_rowuid}{number or string: the specific the unique ID that corresponds to the row to delete}

\item{key_column}{string: the name of the column with the unique ID}
}
\value{
TRUE for successful deletion
}
\description{
Deletes a row or rows by specifying the unique ids for the rows to delete
Note: not implemented in a UI button, but here for administrator use
}
\examples{
\dontrun{
con <- pool::dbPool(DBI::dbConnect(RSQLite::SQLite(), 'iris.db'))
iris <- dplyr::mutate(iris, unique_id = paste0('uid_',dplyr::row_number()))
crudr::cdr_create_tbls_in_db(db_conn_pool = con, db_tbl = iris)
print(here::here('iris.db'))
crudr::cdr_delete_row_in_db(
  db_conn_pool = con,
  db_tbl_name = 'iris', #'iris_deltas'
  value_rowuid = 'uid_1',
  key_column = 'unique_id'
  )
pool::poolClose(con)
}

}
