% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_id.R
\name{cdr_id}
\alias{cdr_id}
\title{A drop-in replacement for \code{\link[DBI:Id]{DBI::Id()}}, but with ordered output}
\usage{
cdr_id(..., table = NULL, schema = NULL, catalog = NULL, cluster = NULL)
}
\arguments{
\item{...}{other table index parameters}

\item{table}{table name as string}

\item{schema}{schema name as string}

\item{catalog}{catalog name as string}

\item{cluster}{cluster name as string}
}
\value{
equivalent \code{\link[DBI:Id]{DBI::Id()}} object
}
\description{
A drop-in replacement for \code{\link[DBI:Id]{DBI::Id()}}, but with ordered output
}
\examples{
cdr_id(table = 'my_table', schema = 'a_schema', some_new_spec = 'in_case_you_need_it')
cdr_id(table = 'table_last', schema = 'schema_3rd', cluster = 'clus_1st', catalog = 'cat_2nd')

}
