% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_js_edit_ctrl.R
\name{cdr_js_edit_ctrl}
\alias{cdr_js_edit_ctrl}
\title{DT Javascript callback to make edit keys
mimic the navigation and cell editing keys of LibreCalc/Excel}
\usage{
cdr_js_edit_ctrl()
}
\value{
javascript for DT::datatable()
}
\description{
DT Javascript callback to make edit keys
mimic the navigation and cell editing keys of LibreCalc/Excel
}
\examples{
cdr_js_edit_ctrl()
}
