% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_make_db_tbls.R
\name{cdr_make_db_tbls}
\alias{cdr_make_db_tbls}
\title{Creates primary and deltas tables in a db}
\usage{
cdr_make_db_tbls(conn_pool, db_tbl, key_field = NULL)
}
\arguments{
\item{conn_pool}{a database connection of class 'pool'}

\item{db_tbl}{a dataframe; the primary table to place into the database}

\item{key_field}{string: a unique ID column name in the data. If Null, creates a proxy}
}
\description{
creates an initial primary table into the database specified by
conn_pool; as well as the corresponding deltas metadata table for
the change tracking
}
\details{
Note: if you want to remove your table, these functions can be helpful:
pool::dbListTables(conn_pool)
pool::dbRemoveTable(conn_pool, 'some_table')
}
