% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_make_unique_ids.R
\name{cdr_make_unique_ids}
\alias{cdr_make_unique_ids}
\title{Title}
\usage{
cdr_make_unique_ids(db_tbl, key_field = "UID")
}
\arguments{
\item{db_tbl}{a dataframe}

\item{key_field}{the name of the column with the Unique ID}
}
\value{
a table with a column of Unique IDs
}
\description{
Title
}
\examples{
crudr::cdr_make_unique_ids(utils::head(iris))
}
