% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_name_delta_tbl.R
\name{cdr_name_delta_tbl}
\alias{cdr_name_delta_tbl}
\title{creates the name of the change tracking table so it's set in one location}
\usage{
cdr_name_delta_tbl(db_tbl_name)
}
\arguments{
\item{db_tbl_name}{char string: name of the primary table the deltas table will track}
}
\value{
char string: name of the table to track changes to the primary table
}
\description{
creates the name of the change tracking table so it's set in one location
}
\examples{
\dontrun{
crudr::cdr_name_delta_tbl('some_table_name')
}
}
