% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdr_update_db_deltas_tbl.R
\name{cdr_update_db_deltas_tbl}
\alias{cdr_update_db_deltas_tbl}
\title{Populates deltas table when user updates primary table}
\usage{
cdr_update_db_deltas_tbl(db_conn_pool, db_tbl_name, to_deltas_tbl)
}
\arguments{
\item{db_conn_pool}{pool connection object: the pool of connections established by the session}

\item{db_tbl_name}{string: name of the specific table the value to update is located in}

\item{to_deltas_tbl}{tibble of delta values to append
Note: to_deltas_tbl$WHEN_EDITED must be a string in timezone 'UTC'}
}
\value{
an in-memory copy of the data just appended to the tracking table
}
\description{
Connects to and posts who-what-when data to a change tracking table
Function is called when a user updates the data in a primary table
}
